+++ src/spline2/MultiBspline1D.hpp
@@ -18,11 +18,96 @@
  * Function signatures modified anticipating its use by a class that can perform data parallel execution
  * - evaluate(...., int first, int last)
  */
-#ifndef QMCPLUSPLUS_MULTIEINSPLINE_1D_ENGINE_HPP
-#define QMCPLUSPLUS_MULTIEINSPLINE_1D_ENGINE_HPP
+
+#ifndef QMCPLUSPLUS_MULTIEINSPLINE_1D_HPP
+#define QMCPLUSPLUS_MULTIEINSPLINE_1D_HPP
 
 namespace qmcplusplus
 {
+
+  template<typename T>
+    struct MultiBspline1D
+    {
+
+      ///define the einsplie object type
+      using spliner_type=typename bspline_traits<T,1>::SplineType;
+      ///define the real type
+      using real_type=typename bspline_traits<T,1>::real_type;
+      ///actual einspline multi-bspline object
+      spliner_type spline_m;
+      ///use allocator
+      //einspline::BsplineAllocator myAllocator;
+
+      MultiBspline1D()
+      {
+        spline_m.coefs=nullptr;
+        spline_m.num_splines=0;
+        spline_m.coefs_size=0;
+      }
+
+      /** create the einspline as used in the builder
+       */
+      template<typename GT, typename BCT>
+      void create(GT& grid, BCT& bc, int num_splines)
+      {
+        if(getAlignedSize<T>(num_splines)!=num_splines)
+          throw std::runtime_error("When creating the data space of MultiBspline1D, num_splines must be padded!\n");
+        spliner_type* temp_spline;
+        temp_spline=einspline::create(temp_spline, grid, bc, num_splines);
+        spline_m=*temp_spline;
+        free(temp_spline);
+      }
+
+      void flush_zero() const
+      {
+        if(spline_m.coefs!=nullptr) std::fill(spline_m.coefs, spline_m.coefs+spline_m.coefs_size, T(0));
+      }
+
+      int num_splines() const
+      {
+        return spline_m.num_splines;
+      }
+
+      size_t sizeInByte() const
+      {
+        return (spline_m.coefs==nullptr)?0:spline_m.coefs_size*sizeof(T);
+      }
+
+      /** copy a single spline to the big table
+       * @param aSpline UBspline_3d_(d,s)
+       * @param int index of aSpline
+       * @param offset_ starting index for the case of multiple domains
+       * @param base_ number of bases
+       */
+      template<typename SingleSpline>
+      void copy_spline(SingleSpline* aSpline,int i, const int offset_, const int base_)
+      {
+        einspline::set(&spline_m,i,aSpline,offset_,base_);
+      }
+
+      template<typename PT, typename VT>
+        inline void evaluate(const PT& r, VT& psi) const
+        {
+          evaluate_v_impl(r,psi.data());
+        }
+
+      template<typename PT, typename VT, typename GT, typename LT>
+        inline void evaluate_vgl(const PT& r, VT& psi, GT& grad, LT& lap) const
+        {
+          evaluate_vgl_impl(r,psi.data(),grad.data(),lap.data());
+        }
+
+      //template<typename PT, typename VT, typename GT, typename HT>
+      //  inline void evaluate_vgh(const PT& r, VT& psi, GT& grad, HT& hess)
+      //  {
+      //    evaluate_vgh_impl(r,psi.data(),grad.data(),hess.data());
+      //  }
+
+      /// compute values only.
+      void evaluate_v_impl(T r, T* restrict vals) const;
+      /// compute VGL.
+      void evaluate_vgl_impl(T r, T* restrict vals, T* restrict grads, T* restrict lapl) const;
+    };
 
   template<typename T>
     inline void
